refactor: expose app module
